+++ test/blackbox-tests/test-cases/ocamldep-multi-stanzas/run.t
@@ -9,26 +9,14 @@
   instead.
   Note: You can use "dune upgrade" to convert your project to dune.
   File "jbuild", line 1, characters 0-0:
-  Warning: Module "Lib" is used in several stanzas:
+  Error: Module "Lib" is used in several stanzas:
   - jbuild:2
   - jbuild:6
   To remove this warning, you must specify an explicit "modules" field in every
   library, executable, and executables stanzas in this jbuild file. Note that
   each module cannot appear in more than one "modules" field - it must belong
   to a single library or executable.
-  Entering directory 'jbuild'
-      ocamldep .lib.objs/lib.ml.d
-        ocamlc .lib.objs/byte/lib.{cmi,cmo,cmt}
-      ocamlopt .lib.objs/native/lib.{cmx,o}
-      ocamlopt lib.{a,cmxa}
-      ocamldep .test.eobjs/lib.ml.d
-      ocamldep .test.eobjs/test.ml.d
-        ocamlc .test.eobjs/byte/lib.{cmi,cmo,cmt}
-        ocamlc .test.eobjs/byte/test.{cmi,cmo,cmt}
-      ocamlopt .test.eobjs/native/lib.{cmx,o}
-      ocamlopt .test.eobjs/native/test.{cmx,o}
-      ocamlopt test.exe
-  foo bar
+  [1]
 
   $ dune build src/a.cma --debug-dep --display short --root jbuild
   Entering directory 'jbuild'
@@ -41,17 +29,14 @@
   instead.
   Note: You can use "dune upgrade" to convert your project to dune.
   File "src/jbuild", line 1, characters 0-0:
-  Warning: Module "X" is used in several stanzas:
+  Error: Module "X" is used in several stanzas:
   - src/jbuild:1
   - src/jbuild:2
   To remove this warning, you must specify an explicit "modules" field in every
   library, executable, and executables stanzas in this jbuild file. Note that
   each module cannot appear in more than one "modules" field - it must belong
   to a single library or executable.
-        ocamlc src/.a.objs/byte/a.{cmi,cmo,cmt}
-      ocamldep src/.a.objs/x.ml.d
-        ocamlc src/.a.objs/byte/a__X.{cmi,cmo,cmt}
-        ocamlc src/a.cma
+  [1]
 
   $ dune exec ./test.exe --debug-dep --display short --root dune
   Entering directory 'dune'