+++ test/blackbox-tests/test-cases/vlib/run.t
@@ -60,7 +60,7 @@ virtual libraries may not implement their virtual modules
   1 | (library
   2 |  (name foo)
   3 |  (private_modules m))
-  Warning: Some modules don't have an implementation.
+  Error: Some modules don't have an implementation.
   You need to add the following field to this stanza:
   
     (modules_without_implementation m)
@@ -111,7 +111,7 @@ virtual libraries may not implement their virtual modules
   File "dune", line 1, characters 0-21:
   1 | (library
   2 |  (name foo))
-  Warning: Some modules don't have an implementation.
+  Error: Some modules don't have an implementation.
   You need to add the following field to this stanza:
   
     (modules_without_implementation m)